Decomposed logic for product suggestions to useProductSuggestions hook.
