+++ src/hooks/index.ts
@@ -13,3 +13,4 @@ export * from './useProductsSort';
 export * from './usePagination';
 export * from './useProductToolbar';
 export * from './useCarousel';
+export * from './useProductSuggestions';
